+++ resources/assets/js_owl/bootstrap.js
@@ -11,11 +11,13 @@ window.axios = require('axios');
 window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
 
 //
+window.Trix = require('trix');
+
 require('./libs/select2');
 require('./libs/noty');
 
 require('./libs/jquery');
-require('./libs/jquery-form');
+// require('./libs/jquery-form');
 require('./libs/moment');
 require('./libs/bootstrap');
 require('./libs/i18next');
@@ -93,6 +95,7 @@ require('./admin/localstorage/tabs');
  require('./admin/form/files');
  require('./admin/form/image');
  require('./admin/form/images');
+ // require('./components/trix');
 
 
 //OLD need test